Refactor elasticsearch logstash client methods
