+++ internal/elasticsearch/logstash/pipeline.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/elastic/terraform-provider-elasticstack/internal/clients"
+	"github.com/elastic/terraform-provider-elasticstack/internal/clients/elasticsearch"
 	"github.com/elastic/terraform-provider-elasticstack/internal/models"
 	"github.com/elastic/terraform-provider-elasticstack/internal/utils"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
@@ -213,7 +214,7 @@ func resourceLogstashPipelinePut(ctx context.Context, d *schema.ResourceData, me
 
 	logstashPipeline.Username = d.Get("username").(string)
 
-	if diags := client.PutLogstashPipeline(ctx, &logstashPipeline); diags.HasError() {
+	if diags := elasticsearch.PutLogstashPipeline(ctx, client, &logstashPipeline); diags.HasError() {
 		return diags
 	}
 
@@ -231,7 +232,7 @@ func resourceLogstashPipelineRead(ctx context.Context, d *schema.ResourceData, m
 		return diags
 	}
 
-	logstashPipeline, diags := client.GetLogstashPipeline(ctx, resourceID)
+	logstashPipeline, diags := elasticsearch.GetLogstashPipeline(ctx, client, resourceID)
 	if logstashPipeline == nil && diags == nil {
 		tflog.Warn(ctx, fmt.Sprintf(`Logstash pipeline "%s" not found, removing from state`, resourceID))
 		d.SetId("")
@@ -289,7 +290,7 @@ func resourceLogstashPipelineDelete(ctx context.Context, d *schema.ResourceData,
 		return diags
 	}
 
-	if diags := client.DeleteLogstashPipeline(ctx, resourceID); diags.HasError() {
+	if diags := elasticsearch.DeleteLogstashPipeline(ctx, client, resourceID); diags.HasError() {
 		return diags
 	}
 	return nil
